To design and implement the "Game of Life" program
Problem Specification:
Life of game includes a two-dimensional rectangular world, each box in the world living with alive or dead cells. The life and death of a cell at the next moment depends on the number of living or dead cells in the adjacent eight squares. If the number of living cells in the adjacent grid is too large, the cell will die at the next moment because of a shortage of resources. On the contrary, if the number of living cells is too small, the cell will die too lonely.
When the current cell is in a dead state, when there are three viable cells in the periphery, the cell becomes a living state. (reproduction)
When the current cell is alive, the cell becomes dead when less than 2 (not including 2) viable cells surround it. (population sparse)
When the current cell is alive, when there are two or three living cells in the vicinity, the cell remains intact.
When the current cell is in a living state, when there are more than three living cells in the vicinity, the cell becomes a dead state. (overcrowding)

The initial cell structure can be defined as a seed, and when all the cells in the seed are treated by the above rule, the first generation cell map can be obtained. Continue to let the rules of the current cell map, you can get the next generation of cell map.

The objectives of this homework are:
1.design and implement a program in C/C++ or Fortran to simulate the "Game of Life"
2.test the program for functionality and correctness
3.measure the performance of the program and optimize the program to improve performance
The program must take the size of the board and the maximum number of generations as command-line arguments. The game will end when there is no change in the board between two generations or when the maximum generation count is reached.(From Spring 2017: CS 432/632/732 Parallel Computing Homework-1)

Program Design:
I want to create two 2-dimensional arrays to represent all the cells in the map, one to represent the current state of life(alive 1 or dead 0) and the other to represent the life of the next generation.
map[row][col] = newmap[row][col]

Life game rules can be simplified, and I will use of SWITCH/CASE comparison can be used to implement the program:
When the number of neighbors is 0, 1, 4, 5, 6, 7, 8, the next state of the cell is death.
When the number of neighbors is 2, the next state of the cell is resurrection.
When the number of neighbors is 3, the next state of the cell is stable.
So,I believed that I don’t need the size (N+2) X (N+2). If I want to create an extra set of cells around the board, I can use MAXCOL+1 to directly create the board. It will output by one FOR LOOP. Then I used the switch and case to count alive neighbors. Because the next-generation state is determined by counting the number of present neighbors.

First,I can use static arrays to input design in program.
static int MAXROW=5;
static int MAXCOL=5;
static int map[5][5];
static int newmap[5][5];
Next, I take the size of the board, initialize cells and the maximum number of generations as command-line arguments. So, I can enter these when compiling the program:
$ gcc log2.c -o log2
$ ./log2 life_num generation_num  map_row map_col//For example, ./log2 12 5 6 7 means 12 initial cells, 5 max generation, row==6, col==7.
Entered the number of alive cells, which will be generated on the map at random.
row=rand()%MAXROW;
col=rand()%MAXROW;
The alive and dead condition of cells is output by two FOR LOOP.
 

Testing Plan:
1.For special cells test: Boundary cells may cause the problem due to their special positions. So, I need to create a lot of basic tests to guarantee that the boundary cells were generated successful. 
2.FOR LOOP: Due to improper use of the FOR LOOP, It may appear dead cycle and so on.
3.Generated time could illegal.
4.The commend was used undeclared.



Test Cases:
Test Case #	Problem Size	Max. Generations	Time Taken
1	                1000x1000	1000                     	91.146937 seconds
2	                 5000x5000	1000                   	2269.373879 seconds
3	                 5000x5000	5000                  	11354.991012 seconds
4	                 10000x10000	1000	                        9078.495322 seconds
5	                 10000x10000	10000	                90833.104832 seconds